requirements.txt
# Weather app Dependencies
# Generated by AI-Native Flask Template

# Web Framework
flask>=3.0.0

# Development & Testing
pytest>=7.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: Add more dependencies as needed
# For database: sqlalchemy>=2.0.0
# For async: asyncio
# For data: pandas>=2.0.0
# For ML: scikit-learn>=1.3.0
